+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 1. The table that is written to within QuestDB can now be specified with `--questdb-table`.
 1. A timeout for QuestDB connections can now be specified with `--questdb-timeout`.
+1. Support for storing exchange rates in a SQLite3 database with `--sqlite3-*`.
 
 ### Changed